requirements.txt
python-telegram-bot==13.7
django==3.2.6
djangorestframework==3.12.4
psycopg2==2.9.1
djangorestframework-simplejwt==4.8.0